refactored Change and ChangeRunner clasess to a proper module
